Tests/LibWeb/Text/expected/css/style-invalidation/viewport-resize-animation-root-font-media.txt

initial animated left: 10px
resized animated left: 20px
resized full invalidations: 0
resized back animated left: 10px
